Migrate kernel API data models into core.
