+++ numba_dpex/core/datamodel/models.py
@@ -8,6 +8,16 @@ from numba.core.datamodel.models import PrimitiveModel, StructModel
 from numba.core.extending import register_model
 
 from numba_dpex.core.exceptions import UnreachableError
+from numba_dpex.core.types.kernel_api.atomic_ref import AtomicRefType
+from numba_dpex.core.types.kernel_api.index_space_ids import (
+    GroupType,
+    ItemType,
+    NdItemType,
+)
+from numba_dpex.core.types.kernel_api.local_accessor import (
+    DpctlMDLocalAccessorType,
+    LocalAccessorType,
+)
 from numba_dpex.utils import address_space
 
 from ..types import (
@@ -250,6 +260,61 @@ class NdRangeModel(StructModel):
         return get_flattened_member_count(self)
 
 
+class AtomicRefModel(StructModel):
+    """Data model for AtomicRefType."""
+
+    def __init__(self, dmm, fe_type):
+        members = [
+            (
+                "ref",
+                types.CPointer(fe_type.dtype, addrspace=fe_type.address_space),
+            ),
+        ]
+        super().__init__(dmm, fe_type, members)
+
+
+class EmptyStructModel(StructModel):
+    """Data model that does not take space. Intended to be used with types that
+    are presented only at typing stage and not represented physically."""
+
+    def __init__(self, dmm, fe_type):
+        members = []
+        super().__init__(dmm, fe_type, members)
+
+
+class DpctlMDLocalAccessorModel(StructModel):
+    """Data model to represent DpctlMDLocalAccessorType.
+
+    Must be the same structure as
+    dpctl/syclinterface/dpctl_sycl_queue_interface.h::MDLocalAccessor.
+
+    Structure intended to be used only on host side of the kernel call.
+    """
+
+    def __init__(self, dmm, fe_type):
+        members = [
+            ("ndim", types.size_t),
+            ("dpctl_type_id", types.int32),
+            ("dim0", types.size_t),
+            ("dim1", types.size_t),
+            ("dim2", types.size_t),
+        ]
+        super().__init__(dmm, fe_type, members)
+
+
+class LocalAccessorModel(StructModel):
+    """
+    Data model for the LocalAccessor type when used in a host-only function.
+    """
+
+    def __init__(self, dmm, fe_type):
+        ndim = fe_type.ndim
+        members = [
+            ("shape", types.UniTuple(types.intp, ndim)),
+        ]
+        super().__init__(dmm, fe_type, members)
+
+
 def _init_data_model_manager() -> datamodel.DataModelManager:
     """Initializes a data model manager used by the SPRIVTarget.
 
@@ -286,6 +351,21 @@ def _init_data_model_manager() -> datamodel.DataModelManager:
 
     dmm.register(IntEnumLiteral, IntEnumLiteralModel)
 
+    # Register the types and data model in the DpexExpTargetContext
+    dmm.register(AtomicRefType, AtomicRefModel)
+
+    # Register the LocalAccessorType type
+    dmm.register(LocalAccessorType, USMArrayDeviceModel)
+
+    # Register the GroupType type
+    dmm.register(GroupType, EmptyStructModel)
+
+    # Register the ItemType type
+    dmm.register(ItemType, EmptyStructModel)
+
+    # Register the NdItemType type
+    dmm.register(NdItemType, EmptyStructModel)
+
     return dmm
 
 
@@ -311,3 +391,18 @@ register_model(RangeType)(RangeModel)
 
 # Register the NdRangeType type
 register_model(NdRangeType)(NdRangeModel)
+
+# Register the GroupType type
+register_model(GroupType)(EmptyStructModel)
+
+# Register the ItemType type
+register_model(ItemType)(EmptyStructModel)
+
+# Register the NdItemType type
+register_model(NdItemType)(EmptyStructModel)
+
+# Register the MDLocalAccessorType type
+register_model(DpctlMDLocalAccessorType)(DpctlMDLocalAccessorModel)
+
+# Register the LocalAccessorType type
+register_model(LocalAccessorType)(LocalAccessorModel)
